+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project are documented in this file.
 
 
 ## 2025-09-28
+### Added
+- Navigation via `MyDrawer.js`
 ### Changed
 - Aestetic changes to workspace selector in `MyDrawer.js`
 - Added error and loading messages in `MyDrawer.js`